rename files to match their functional components
