+++ client/app/caseDistribution/components/LeverButtonsWrapper.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { LeverCancelButton } from './LeverButtons';
-import { LeverSaveButton } from './LeverModal';
+import { LeverCancelButton } from './LeverCancelButton';
+import { LeverSaveButton } from './LeverSaveButton';
 import { css } from 'glamor';
 
 const saveButtonStyling = css({
